requirements.txt
bcrypt==3.2.0
bottle==0.12.19
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
ecdsa==0.14.1
idna==2.10
JWT-Bottle==2021.1.4
marshmallow==3.10.0
passlib==1.7.4
peewee==3.14.1
pyasn1==0.4.8
pycparser==2.20
pycrypto==2.6.1
PyJWT==2.0.1
python-dotenv==0.15.0
python-jose==3.2.0
requests==2.25.1
rsa==4.7.1
six==1.15.0
urllib3==1.26.3
